Fix CI: Windows build, UBSan alignment, clang-tidy warnings
- xm_jit_debug.c: wrap POSIX signal handler code with #ifndef _WIN32;
  provide empty stub on Windows (SEH implementation deferred).
- xi_analysis.c: replace VLA (not supported by MSVC) with heap
  allocation hoisted before convergence loop.
- xworker.c: use xr_malloc_aligned(XR_CACHE_LINE) for XrWorker array
  to satisfy _Alignas(64) in XrProcStats (fixes UBSan misaligned
  access).
- xtype.h: merge identical XR_KIND_INT/BOOL/NULL cases to eliminate
  bugprone-branch-clone warning.
- xparse_import.c: explicit cast strlen() to int to silence
  bugprone-narrowing-conversions.
- xunicode.c: add trailing comments to array entries for consistent
  clang-format-19 alignment.

+++ src/frontend/parser/xparse_import.c
@@ -64,7 +64,7 @@ static void parse_unquoted_module(Parser *parser, char **out_name, ImportType *o
     if (xr_parser_match(parser, TK_SLASH)) {
         xr_parser_consume(parser, TK_NAME, "expected package name");
         int name_len = parser->previous.length;
-        int total_len = strlen(first_part) + 1 + name_len;
+        int total_len = (int) strlen(first_part) + 1 + name_len;
         *out_name = (char *) ast_alloc(parser->X, (size_t) total_len + 1);
         snprintf(*out_name, total_len + 1, "%s/%.*s", first_part, name_len, parser->previous.start);
         *out_type = IMPORT_PACKAGE;
